requirements.txt
numpy
scikit-learn
sentence-transformers
faiss-cpu
pydantic
fastapi
uvicorn
python-dotenv
tiktoken
redis
streamlit
